requirements.txt
PySide6>=6.6.0
numpy>=1.26.0
mss>=9.0.1
Pillow>=10.1.0
